# openapi.yaml
openapi: 3.0.0
info:
  title: Employee Management API
  version: 1.0.0
  description: REST API to manage employee records using Golang and MySQL
servers:
  - url: https://p2-gc1-h8-jersbs-4c4beecd5a82.herokuapp.com/
    description: Production server

paths:
  /employees:
    post:
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEmployee'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid input or duplicate email
    get:
      summary: Get all employees
      responses:
        '200':
          description: List of employees (ID, name, email only)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    email:
                      type: string

  /employees/{id}:
    get:
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Employee not found

    put:
      summary: Update an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEmployee'
      responses:
        '200':
          description: Employee updated
        '404':
          description: Employee not found

    delete:
      summary: Delete an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee deleted
        '404':
          description: Employee not found

components:
  schemas:
    NewEmployee:
      type: object
      required:
        - name
        - email
        - phone
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string

    Employee:
      allOf:
        - $ref: '#/components/schemas/NewEmployee'
        - type: object
          properties:
            id:
              type: integer